+++ application/helpers/SurveyRuntimeHelper.php
@@ -287,7 +287,7 @@ class SurveyRuntimeHelper {
             {
                 // then trying to resubmit (e.g. Next, Previous, Submit) from a cached copy of the page
                 // Does not try to save anything from the page to the database
-                $moveResult = LimeExpressionManager::GetLastMoveResult(true);
+                $moveResult = LimeExpressionManager::JumpTo($_SESSION[$LEMsessid]['step'], false, false, true);// We JumpTo current step : see bug #11404
                 if (isset($_POST['thisstep']) && isset($moveResult['seq']) && $_POST['thisstep'] == $moveResult['seq'])
                 {
                     // then pressing F5 or otherwise refreshing the current page, which is OK
@@ -302,6 +302,7 @@ class SurveyRuntimeHelper {
                     $invalidLastPage=true;
                     $backpopup=gT("Please use the LimeSurvey navigation buttons or index.  It appears you attempted to use the browser back button to re-submit a page.");
                 }
+
             }
             if(isset($move) && $move=="clearcancel")
             {