+++ CHANGELOG.md
@@ -1,27 +1,36 @@
 # Change Log
 Only partial changelog, [commit history](https://framagit.org/Shnoulle/LimeSurvey/commits/2.06_SondagesPro) show all changelog.
 
-## [not releazed]
+## [1.0.21]] - 2016-06-17
+
+### Feature
+- Allow to set column number on ranking question
+
 ### Fix
 - EM tips are empty after reloading page via browser (F5)
 - Ranking question : Alert are not show every time
 - Ranking question : Add answers can broke Survey DB
 
 ## [1.0.20] - 2016-06-16
+
 ### Feature
 - afterSurveyMenuLoad event to add survey specific menu items
+
 ### Fix
 - Improvement on Conditions page with array (number) questions
 - Google Analytics code not running
 
 ## [1.0.19] - 2016-06-09
+
 ### Fix
 - Better loading of plugins for command
 - Default email format in survey to html
 
 ## [1.0.17] - 2016-05-19
+
 ### Feature
 - Add cssclass question attribute
+
 ### Fix
 - Plugin survey setting type "checkbox" does not properly save
 - Map question : google.maps fix for hidden text element
@@ -35,9 +44,11 @@ Only partial changelog, [commit history](https://framagit.org/Shnoulle/LimeSurve
 - New token table firstname/lastname to 150
 
 ## [1.0.9] - 2016-04-22
+
 ### Feature
 - beforeController event plugin (for web)
 - newUnsecureRequest : plugin direct request without CRSF
+
 ### Fix
 - Checked responses are not read when load "surveys uploaded file"
 - Using mktime() EM function broke survey administration
@@ -48,6 +59,7 @@ Only partial changelog, [commit history](https://framagit.org/Shnoulle/LimeSurve
 - thousand separator break slider in some condition
 
 ## [1.0.6] - 2016-03-20
+
 ### Fix
 - Issue with relevance on X scale with multilingual
 - Languages can not be updated in label set administration
@@ -59,9 +71,11 @@ Only partial changelog, [commit history](https://framagit.org/Shnoulle/LimeSurve
 - [Security] issue when saving/loading responses on public survey
 
 ## [1.0.0] - 2016-03-01
+
 ### Fix
 - Higher risk that the emails are rated as Spam
 - Filter script in Plugin management and Survey Logic file.
+
 ### Updated
 - Use updatable from config and use it, set to updatable=false